refactor: automatically create recommended config based on rule config
